unittest/fake_db: Fix issue where field3 was FT_F32 in the schema but uint32_t in the struct.
This led to some weird stuff.

+++ src/libtsdb/unittest/fake_db.h
@@ -9,7 +9,7 @@ struct data_point
 {
     uint32_t    field1;
     double      field2;
-    uint32_t    field3;
+    float       field3;
     bool        is_non_null[3];
 };
